requirements.txt
Django==1.10
djangorestframework==3.4.6
Markdown==2.6.6
psycopg2==2.6.2
django-filter
